requirements.txt
# Base Flask
Flask==3.1.1
Flask-Cors==6.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Flask-Mail

# LLM y AI
openai
crewai
litellm
langchain

# Dependencias internas de Flask
blinker==1.9.0
click==8.2.1
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3

#Seguridad
bcrypt
PyJWT

#gunicorn
gunicorn